MEDIUM: shctx: use unsigned int for len and block_count
Allows bigger objects to be cached in the shctx, the first
implementation was only storing small ssl session, but we want to store
bigger HTTP response.

+++ include/types/shctx.h
@@ -21,8 +21,8 @@
 /* generic shctx struct */
 struct shared_block {
 	struct list list;
-	short int len;          /* data length for the row */
-	short int block_count;  /* number of blocks */
+	unsigned int len;          /* data length for the row */
+	unsigned int block_count;  /* number of blocks */
 	unsigned int refcount;
 	unsigned char data[0];
 };
